Move app-header/ to common/
