+++ components/common/index.ts
@@ -1,2 +1,3 @@
+export { default as AppHeader } from "./app-header";
 export * from "./card"
 export * from "./expandable-panels";
