+++ pages/conditions-d-utilisation-leximpactpop.jsx
@@ -2,7 +2,7 @@ import { flow } from "lodash";
 import Head from "next/head";
 import { Fragment, PureComponent } from "react";
 
-import AppHeader from "../components/app-header";
+import { AppHeader } from "../components/common";
 import TexteCguLeximpactPop from "../components/presentation-cgu/cgu-leximpact-pop/texte-cgu-leximpact-pop";
 import withRoot from "../lib/withRoot";
